[1.3.5] Merged revision(s) 6122-6126 from trunk: r6122 (Fixed #8044 (ElementFrozen for Players is desynced after spawnPlayer)) r6123 (Fixed server crash during shutdown. Modified server stats labels.) r6124 (Fixed 7702 (exports: Call to non-running client resource (name) [string "?"]) patch from Fabio(GNR)) r6125 (Added warning message for server scripts that might be causing a long freeze. Applied a couple of minor source tweaks.) r6126 (Updated stats output)

+++ branches/1.3.5/MTA10/mods/deathmatch/logic/CPacketHandler.cpp
@@ -4582,7 +4582,7 @@ void CPacketHandler::Packet_ResourceStart ( NetBitStreamInterface& bitStream )
                     bitStream.Read ( (char*)chunkChecksum.mD5, sizeof ( chunkChecksum.mD5 ) );
                     bitStream.Read ( dChunkDataSize );
 
-                    uiTotalSizeProcessed += dChunkDataSize;
+                    uiTotalSizeProcessed += (uint)dChunkDataSize;
                     if ( uiTotalSizeProcessed / 1024 / 1024 > 50 )
                         g_pCore->UpdateDummyProgress( uiTotalSizeProcessed / 1024 / 1024, " MB" );
